[Code] Refactored code
