--- config/supervised/touching_3.yaml
@@ -1,9 +1,0 @@
-dataset_type: "touching"
-train_val_directions: [ 'E', 'N', 'S', 'W', 'C' ]
-train_val_classes: [ 0, 1, 2, 4, 5, 6, 7, 8, 9, 10 ]
-test_directions: [ 'E', 'N', 'S', 'W', 'C' ]
-test_classes: [ 3 ]
-dataset_folder: "/media/mbed/internal/backup/haptr"
-dataset_file: "touching.mat"
-signal_length: 192
-signal_start: 0
