+++ models/__init__.py
@@ -1,2 +1,1 @@
-from .autoencoder import TimeSeriesAutoencoder, TimeSeriesAutoencoderConfig, RCNNAutoencoder, RCNNAutoencoderConfig
-from .clustering import ClusteringModel
+from . import autoencoders, dec
